refactor(frontend): Merge UpdateUserProfileRequestDto into user.d.ts
This commit merges the UpdateUserProfileRequestDto interface directly
into user.d.ts to centralize user-related type definitions.

Key changes:
- Moved UpdateUserProfileRequestDto definition from its dedicated file
  into user.d.ts.
- Deleted the now redundant update-user-profile-request.d.ts file.
- Removed the corresponding export from index.ts.
