+++ apps/admin/src/types/index.ts
@@ -19,7 +19,6 @@ export * from './biography.d';
 
 export * from './memberFace.d'; 
 export * from './uploaded-file.d';
-export * from './update-user-profile-request.d';
 
 // NEWLY ADDED PUBLIC-FACING DTO EXPORTS
 export * from './common-dto.d';
